create Event implemented

+++ api/scheduler/src/main/java/com/example/scheduler/entities/ParticipantsEntityPK.java
@@ -9,27 +9,27 @@ public class ParticipantsEntityPK implements Serializable {
 
     @Id
     @Column(name = "event_id")
-    private int eventId;
+    private Long eventId;
 
     @Id
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
 
-    public int getEventId() {
+    public Long getEventId() {
         return eventId;
     }
 
-    public void setEventId(int eventId) {
+    public void setEventId(Long eventId) {
         this.eventId = eventId;
     }
 
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
